Update to 3.10.0
In the last commit, I switched the checksum from SHA-1 to SHA2-512
because that was all that was available on the download index [1]; this
time, only SHA2-512 is available again, so let’s switch the checksum
back again. (I have no idea what the Apache folks are doing there, to be
honest. Whatever.)

[1]: https://www.apache.org/dist/jena/binaries/