updpkg: 0.93.1 + cleanup + hack
I had the unpleasant surprise to find that something changed in the Qt5
libraries shipped in 0.93.1, in such a way that they're not quite
compatible with the system ones on Arch. Fortunately I found a way
around that, which is frankly horrifying, but as long as you don't look
at it it's probably alright!

On the plus side, I've made some cleanups:
- Drop the 32-bit package (like in xnviewmp).
- Drop the trash workaround, it's fixed in 0.93.1.
- No need to special-case the "styles" folder any more, since it's now
  shipped as well.