libpurple-lurch-git: fix submodule issue
The upstream `.gitmodules` file specifies relative paths that can be
correctly resolved in a direct clone of the upstream GitHub repository,
but not in the local filesystem clone that makepkg does.

Thanks to eskimod <https://aur.archlinux.org/account/eskimod> for an elegant solution.

The `submodule sync` is skipped, as it should happen during the `init`
anyway and might overwrite local changes someone might want to add.