Fix build, actually run tests, and do some cleanup
We depend on python-sphinx again (I don't recall why I changed it). We
no longer run automake and autoconf as specified in the wslay README
because they are redundant. checkdepends is used instead of optdepends,
and the check() function is now implemented. A redundant message was
removed.

Thanks to J5lx for the suggestions.