binfmt: make parser more robust
Make sure we no longer match on the emulation binary, but rather on the
entry-name. The emulation binary might be a different one (see i386 and
i486 for instance). Furthermore, make sure to skip empty entries.

Lastly, note that upstream qemu now has a script to do the exact same
thing. We should switch some day..

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>