Add patch for COE file output
COE output (`srec_cat some_file --binary -o test.coe --coe`) would always fail with:

```
The Coefficient (.COE) Files (Xilinx) output format
    is unable to cope with holes in the data,however there is a hole at
    0x0000..0x0017.
```

Even if the input was a plain binary file, which can't have any holes.
The cause was that the internal address value of the COE output was
never incremented, and thus got out of sync with the input address,
looking like a hole. Simple one-line fix.

COE output was added in 1.61, which has the same problem - apparently it
was never tested?!

Signed-off-by: vicencb <vicencb@gmail.com>