upgpkg: hmcl 3.2.129-4
For now release version 3.2.129 of hmcl from upstream might not be
built other than use open source implements "openjdk 8".
When launch it, there is not need to hard-coded the java-runtime
with only version 8.

From [upstream]<huanghongxun/HMCL#509>, implementation of the plan for java version 11 might
be accomplished in the not too distant future.