Dependency cleanup (see below).
Dependencies have been trimmed down, particularly PostgreSQL (PKGBUILD
now only relies on postgresql-libs) and a few others. llvm has been
moved to `makedepends` and Redis has been added to `optdepends`. Redis
is still required for sentry-celery to function, but because Sentry can
be installed on a server different from that hosting Redis, I've
relented and moved it into `optdepends`.

I'm not sure this is the best place to move Redis since it's not an
optional dependency, but I also don't want Sentry to have a hard
dependency on Redis in circumstances where a server or container might
be spun up specifically for Sentry support when a Redis cluster resides
elsewhere.

If this causes issues with other users, I may revert the change. I'm
hoping to leave it as is, because if you really need Sentry installed,
you probably ought to know a bit about what its requirements are for
self-hosting the on-premise version.

Minio has also been mentioned in the optdepends.