Add graal-native-image to optdepends and bump pkgrel
The native-image plugin used to be part of this package, but Oracle
extracted it into a separate component with the 19.0.0 release (which I
failed to notice when first creating the release). We follow suit by
setting up a separate AUR package for it, graal-native-image(-bin), but
recommend that people install that via optdepends.

The pkgrel is bumped so that people will see the new dependency as soon
as possible. (I think that should work, though I can’t very well try it
out before pushing this commit…)