Use repo for downloading correct versions of dependencies
Previously, sources were downloaded by manually specifying branches for
projects in the sources array in the PKGBUILD.

This patch follows the instructions in the AMDVLK README to use the
"repo" tool for managing dependencies.

This is needed to avoid continually breaking the build by pulling the
wrong versions of dependencies. The AMDVLK project maintains a
"default.xml" file describing the required versions of each of the other
repositories (but oddly hasn't yet specified `wsa`, so there's a
temporary fix with the master branch of that still being in the sources
array).

v2: Fix pkgver sourcing, and use release tag from AMDVLK if available