split configuration and services for server and client
We split the single service and config file into one each for the
server and client. This is adapted from, and is more in line with, the
non-bin package and should ease users' switching between the two
versions of the package.

Please note that the name of the systemd service has changed due to
the split, and that you will probably need to review/adapt your
configuration files.