Add DDNet Server systemd service, drop desktop
Instead of creating a .desktop file for the DDNet-Server executable,
use a systemd unit file which runs the server instance as a user
called "ddnet" and the service's data in "/var/lib/ddnet".

ddnet-sysusers.conf is responsible for creating the user and group
if they don't exist. It won't remove them upon package removal
though, but one can use 'userdel ddnet' for that.

ddnet-tmpfiles.conf is responsible for creating /var/lib/ddnet with
the user and group newly created by sysusers.d config file. There
will be stored autoexec_server.log, a "records" directory and any
other persistent data.

When starting the systemd unit file "ddnet-server.service", it runs
the DDNet-Server executable with work directory "/var/lib/ddnet" as
user "ddnet", which allows to store the persistent files in there.

As a consequence of this change, "imagemagick" and "gendesk" are no
longer needed for creating .png icon and .desktop files, so they
were removed from makedepends array.

.gitignore now includes *.conf and *.service to the exception list