Remove pkgver() function entirely
The pkgver() function is great if the PKGBUILD is tracking a rolling
VCS, but this PKGBUILD follows the latest stable tag.  We already update
the pkgver variable and file checksums on every update so AUR helpers
can update automatically and show the real version you'll update to in
their GUI/text output.