Install font as OTB file
This is due to the recent changes in Pango 1.44:

https://gitlab.gnome.org/GNOME/pango/issues/386

Sadly, you can no longer use the font in traditional applications like
XTerm. It's also NOT possible to install the font in both formats at the
same time, since Pango will pick up the wrong one -- maybe that will get
fixed in later versions.