Fix most illegal instructions when shipping asterisk.
While this works in AUR, because users are building the PKGBUILD,
distributions wanting to use this PKGBUILD as-is and ship packages
produced with it will result in illegal instructions at runtime.

Without that fix, if you build this packages on a machine with
a processor that supports avx (Advanced Vector Extensions),
libpjsip.so.2 ended up with vpxor instructions that were also
executed on machines with a processor that did not have such
extension, which resulted in an illegal instruction at asterisk
startup if the machine running it didn't have AVX.

Though for some reason, with this patch, building asterisk on a
machine with avx still results in res_pjsip_send_to_voicemail.so
having one vpxor instruction.

Hopefully even with that, asterisk can still start on machines
that do not have AVX.

To fix most illegal instructions:
- Native builds were disabled by using the ChangeLog documentation:
  "Those who need different -march= values, please, go for
   ./configure
   make menuselect.makeopts or make menuselect
  ./menuselect/menuselect --disable BUILD_NATIVE"
- The build configuration was changed to use an external
  pjproject library. The AUR PKGBUILD for pjproject also
  had several fixes to prevent it from causing illegal
  instructions.

On x86_64 I verified that most asterisk ELF files did not
have vpxor instructions anymore with:
  $ pacman -Q -l asterisk | \
                 awk '{print $2}' | xargs file | grep ELF | \
                 sed 's#:.*##' | xargs objdump -D | grep vpxor

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>