Patch to fix go-pie related chroot issues
This patch (proposed upstream and likely to be in the next release)
fixes issues with go-pie. Arch systems may have either go or go-pie by
default, go works, but go-pie does not allow both the chroot and
user switch to drop root permissions. Without this go-pie compiled
gitlab-pages would only work if run as root with no uid switching.

Note this also solves the previously patched dynamic linking issue, so
drops that monkey patch in favor of this one.