Warn about Yad menu bug in versions 1.0 to 4.1
A bug was introduced into Yad 1.0 whereby right clicking on the tray
icon to get a context menu would throw a GTK error instead of actually
opening the menu as you'd expect.

This was not an issue in 0.42.x or earlier, and has been fixed on master
but isn't currently released. The next version including the fix is
likely to be 4.2, but since their version scheme changed recently it's
difficult to tell. If there are other changes made it could be 5.0.

Either way, when the next version is released the message will go away.