Remove --enable-omx-rpi
Although this option is specific to the Raspberry Pi, it was
compiling fine so it was ok to leave it there.

But a recent upstream ffmpeg commit[1] broke build with this
option, making it to fail with the following error:
‘OMX_IndexConfigBrcmVideoRequestIFrame’ undeclared.

Now it needs headers[1] that are specific to the Raspeberry Pi
firmware, so it will be removed.

References
----------
[1] https://git.ffmpeg.org/gitweb/ffmpeg.git/commitdiff/76c82843ccad11666603e3f1f6d9d9a36c780a32
[2] https://github.com/raspberrypi/firmware/blob/1.20190819/hardfp/opt/vc/include/IL/OMX_Index.h#L381