Fix source download with curl 7.66.0
The curl update 7.65.3->7.66.0 breaks download with the
following message:
Warning: Binary output can mess up your terminal. Use "--output -" to tell
Warning: curl to output it to your terminal anyway, or consider "--output
Warning: <FILE>" to save to a file.

Then it started the download and even acquired the correct
total filesize (meaning that the huge download url, the one
with a temporary string, was correctly acquired), but then
it downloads the source file with zero size.

I'm not sure where this is coming from since we were already
specifying '-o %o' (--output <FILE>) in the outter curl
command, and the inner curl command was not downloading any
binary data (it was just getting the download url). I'm
mostly sure that this is comming from some makepkg internals
because the error does not happen when executing the same
curl commands in the terminal.

Redefined the download method and it is now downloading fine. The
strange redirection message that makepkg was giving is also gone.
Now we first get the download url separately and use it later in
the DLAGENTS command, instead of doing everything at the same time
inside DLAGENTS. I hope this continues to work.