Make extraction verbose, instead of installation
I want to see which certificates are provided by upstream,
but it is not necessary to know its installation path as
the pathis only one.

So, I removed the quiet flag from unzip to make it print
extracted files, and I remove verbose flag from install
command to make it silence.