Update to v1.2.15
Updates the package to v1.2.15, adds skia build, puts maintainer names
in PKGBUILD. Skia deps are downloaded manually as downloading them
during prepare turns out to be bad practice, and long options and
verbosity are used for clarity, though it's not common or necessary.
EULA message was removed as there's no need for it and makepkg should
run without user input anyways.