Responder now requests root privileges on startup
Before this change, Responder.py would simply quit if not given root privileges
and state that it needed to be run as root. Now it requests root privileges,
and if they are not given, it quits as before. Note that this does not apply
when running responder with "--version" or "--help" ("-h" included), as those
options do not require root.