PKGBUILD: Follow Arch Linux practices for pkgver
The kernel provided by Clear Linux is, for example, 5.3.9-863. The
important part of that is the upstream Linux kernel version, which
is followed by Clear's own patch level. It is unfortunate that the
patch level uses a dash because this makes it difficult to indicate
revisions to the Arch Linux package itself.

I am changing this to follow Arch Linux best practices. The version
seen by pacman will be the upstream version followed by the package
revision. The version reported by `uname -sr` will stay the same as
it is currently, ie: Linux 5.3.9-863.native.