Initial commit
With the 19.3.0 release, GraalVM supports two Java versions: Java 8 and
Java 11. Accordingly, the former graal-bin package is split into two:
jdk8-graalvm-bin and jdk11-graalvm-bin (this package). We use this
opportunity to make several other adjustments as well.

The main product name seems to be “GraalVM”, not just “Graal”, so we
reflect this in the package name and in the Java name.

We no longer provide a version of the package name without the “-bin”
suffix. It doesn’t seem likely that a built-from-source version will be
added to the AUR soon, and other packages like jdk8-openj9-bin or
jdk8-j9-bin don’t provide non-bin versions either.

We no longer recommend a native-image package. The old package only did
this because it used to include that functionality; this new package has
never included native-image, so I don’t see the need to recommend it.
People interested in it should be able to find it easily enough.

A test script is included, which I’d been using locally for some time
already. It assumes you’ve locally installed current versions of the
GraalVM, FastR, TruffleRuby, GraalPython and native-image packages, and
is typically run just before pushing the updated PKGBUILDs to the AUR.
Note that the jdk11 version of native-image is not yet functional, so
the script currently fails.

And finally, a .gitignore file never hurts.