Initial commit
With the 19.3.0 release, GraalVM supports two Java versions: Java 8 and
Java 11. This also affects all sibling packages, including FastR, so the
former fastr-bin package is split in two: fastr-jdk8-bin (this package)
and fastr-jdk11-bin. We use this opportunity to make several other
adjustments as well.

We no longer provide a version of the package name without the “-bin”
suffix. It doesn’t seem likely that a built-from-source version will be
added to the AUR soon, and other packages like jdk8-openj9-bin or
jdk8-j9-bin don’t provide non-bin versions either.

Since this release no longer includes the GCC libraries required by
FastR, we have to depend on gcc-libs and also gcc6-libs (FastR requires
an older version of libgfortran.so than the one provided by gcc-libs).
gcc6-libs doesn’t install its libs into /usr/lib, so we symlink its
libgfortran.so.3 into FastR’s lib folder; this works for R, but not for
Rscript, so for now the Rscript launcher is slightly tweaked to update
the LD_LIBRARY_PATH. This may change in the future, depending on
feedback from the FastR developers.

And finally, a .gitignore file never hurts.