Initial commit
With the 19.3.0 release, GraalVM supports two Java versions: Java 8 and
Java 11. This also affects all sibling packages, including native-image,
so the former graal-native-image-bin package is split in two:
native-image-jdk8-bin (this package) and native-image-jdk11-bin. We use
this opportunity to make several other adjustments as well.

We remove “graal” from the package name, to balance out the addition of
the Java version (avoiding an overlong package name) and for consistency
with FastR, TruffleRuby and GraalPython, which don’t include “graal” as
a separate component either.

We no longer provide a version of the package name without the “-bin”
suffix. It doesn’t seem likely that a built-from-source version will be
added to the AUR soon, and other packages like jdk8-openj9-bin or
jdk8-j9-bin don’t provide non-bin versions either.

To make it possible to install native-image-jdk8-bin and
native-image-jdk11-bin in parallel, the symlink in /usr/bin is removed
(as it would otherwise conflict between the two packages). Users can add
/usr/lib/jvm/default/bin/ or a similar directory to their $PATH, or
otherwise ensure that they can still run the right native-image without
specifying its full path.

And finally, a .gitignore file never hurts.