Fix version number reported by library.
Distributed uses versioneer to handle version numbering. The tarballs
provided by codeload.github.com were taken straight from the source, in
which case versioneer uses Git to determine a version number. This meant
it was picking up the hash of the latest commit to the AUR package's
repository and giving a version number like '0+untagged.20.g4a90f33' for
both the distributed.__version__ attribute in the library, and the
.egg-info metadata directory.

The tarballs uploaded to PyPI are generated using 'python setup.py
sdist' which makes versioneer embed the correct version number rather
than loading it dynamically. Switching to these as the source for the
package gets us the correct version number in the built library.