Explicitly enable PulseAudio support
Always build support for the PulseAudio volume widget. This is more
robust than relying on Meson automatically deciding whether to enable
it or not because automatic selection can result in a built package
which depends on PulseAudio at runtime, but it has no dependency on
its package, which in turn allows the package manager to remove the
PulseAudio package even if it's needed by wf-shell-git. This change
prevents this issue.