upgpkg: 0.94.2-2 (big refactor)
The latest update to libwebp (1.1) broke XnView, because it tries to
link against the system libwebpmux.so while using the packaged
libwebp.so (in Plugins).

Because of this, I realised that there is no point in using the packaged
plugins, since they are all libraries available in standard Arch
packages. This commit replaces them with symlinks to system libs (as
XnView is still unhappy if it cannot find them in Plugins/), and adds
the relevant dependencies.

While at it, I also realised that a lot of files are junk and don't need
to be installed, and that many permissions are wrong. Fix all that by
specifying which files/directories to install (instead of removing
some), and only making executable those that should be.

Finally, it looks the bug that prevented me from getting rid of lib/
altogether is gone, so there's no need to symlink the Qt plugin
directories any more. As a result, the only use we still have for
xnview.sh is the LD_PRELOAD hack, and we don't need the one included in
the archive any more.