update to 3.3.1
Major overhaul involved.
In the upstream repo are no tags, so the versions are build from the specific commit.
The sysmon module seems to be gone, only vwmterm3 is there. And dunno if that one can be separated in an extra package.
Deps changed significantly: libconfig and protothread are new (need still a check if one of them is makedepends), ncurses and glibc are still a dep, but are pulled via the other deps. libpseudo and libgtop seem to be gone.

Build process needs currently some patching: Like the deps the ncurses import won't work. But also cmake couldn't find protothread.
Still not building, because of libvterm-vwm. Maybe version 9.0 is to recent. This will be clarified soon.