upgpkg: 0.95
libfreetype is now shipped in the archive, but we need to remove it
because it is not compatible with the system libfontconfig.

Additionally, the shipped platformthemes libs don't seem to have been
updated for the new Qt libs (5.12.6), so setting QT_QPA_PLATFORMTHEME
triggers:

    Cannot mix incompatible Qt library (version 0x50907) with this library
    (version 0x50c06)

Not much can be done about this, short of using the system Qt libs
(which is what my xnviewmp-system-libs package does).