Enable Czech and Russian translations
The translations are not compiled by default. Specific ifdefs have to
be enabled to include the localization files.
There's no handy switch in makefile's to pass the defines to gcc.
Fortunately, it can be done via DEFINES environmental variable.
(Setting DEFINES on make command-line doesn't work, because it would
override any changes made to DEFINES in makefiles.)