newpkg: gnu-efi-libs-fedora-git
The name "gnu-efi-libs-fedora-git" has been used before, it was merged
into "gnu-efi-libs-git" by PRQ#3528. I don't think the latter name is
appropriate since it suggests this is the official development version
of gnu-efi, which it is not: the official repository can be found at
https://sourceforge.net/p/gnu-efi/code/, while this package uses a fork
maintained by Red Hat employee Peter Jones (vathpela).