fix missing exclusion on pacman package listing
there is an assumption within the tllocalmgr script that the package
name is of form 'texlive-local', which is used to filter out the list
of tex packages so that we know which system (from archlinux repo) tex
packages are actually installed. With this new package name,
'tllocalmgr-git', this is not excluded leading to a startup error.