restore compilation with Pacman 5.2
Since Pacman version 5.2.0, makepkg exits whenever any command in a
pkgver function returns a non-zero exit code. A recent correction from
the git-makepkg-templates [1] adapts to this, restoring correct
processing of the PKGBUILD.

As this concerns only the build process, not changing any of the
resulting package content, the pkgrel is kept constant, to avoid
triggering any unnecessary rebuild.

[1]: https://github.com/dffischer/git-makepkg-templates/commit/d91fca9b16dd03bc0bf924dbd95dd24b06e232fe