Use open-ended dependencies.
Ruby has the unfortunate habit of being way too strict in its semantic
versioning. In this case, we think it's better to use open-ended
dependencies for now, as most repository packages do.

In the event that stuff breaks, we can reconsider this. For now it seems
to work just fine.

Unfortunately, this does require us to rebuild the gem file from source.
This is slightly more work, but we believe the experience will be better
for everyone in the end.