Make sure to link with Gtk3
Upstream tecmake.mak makefile is kind of weird in guessing when to use
Gtk2 vs Gtk3. It makes the decision upon the Linux kernel version. It
chooses Gtk3 for Linux 4.x. Unfortunately it doesn't know Linux 5.x,
and thus it defaults to Gtk2 for that kernel.

So use USE_GTK3=Yes option to make sure Gtk3 is used unconditionally.