zathura-cb-git: Fix PKGBUILD
As I did a long time ago with the stable release: This depends on libarchive instead of the several separate packer programs.
Additionally fix the build process and sync the [community] PKGBUILD.
The switch to zathura from zathura-git was intentional. The latter provides zathura and as long there is no breaking change, the base program can be choosen freely.