Switch source. This updates to version 0.40.0.20200219.65.
Not sure if the author of this is github source is related to the
the official autotrace source code from sourceforge.  ImageMagick
started to use it[1][2] and now requires autotrace >= 0.31.2 from
this github source. Anyway, this new source is now maintained and
it fixes at lot of issues that we needed to patch on the original
autotrace from sourceforge, which is not being maintained since a
long time.

References
----------
[1] https://github.com/ImageMagick/ImageMagick/issues/1848
[2] https://github.com/ImageMagick/ImageMagick/commit/51968547b98bde1cd6c6cc7145faf3cc1f365a22