Add pkgconf to makedepends
SELinux userspace libraries and tools require pkg-config in order to
build with code using PCRE or DBUS. pkgconf should be installed on build
systems, as it belongs to group base-devel, but some users did not
install this package and report issues such as
https://aur.archlinux.org/packages/libselinux/#comment-731794

Make everyone happy by adding pkgconf to the makedepends array. This
should not modify the build (as pkgconf was already supposed to be
installed), so there is no need to bump the pkgrel.