Don’t try to copy missing file
This file was renamed from 3rd_party_licenses_graalpython.txt to
THIRD_PARTY_LICENSE_GRAALPYTHON.txt with the 20.0.0 release; in the
previous commit, I had added the new name, but not removed the old one.
This didn’t result in an error locally because makepkg doesn’t clean the
srcdir by default, so the file was still there. AUR user florensie [1]
noticed the mistake.

[1]: https://aur.archlinux.org/account/florensie/