Add patch to update url version
As Turtl's codebase continues to be left unmaintained, more and more build
issues appear with it due to changes in Rust. Like with rusqlite, recently the
url library started experiencing build issues with the current Rust compiler.
So I added another patch to update the url version from v1.6.0 to v1.7.2, which
appears to work around the issue. There's still an ever increasing amount of
build warnings, though all of the unit and integration tests still pass, so for
now, its probably okay.

I didn't update it to the latest version because, like with rusqlite, I was
worried about the possibility of API-breaking changes that would require
patching the source code to fix.

The long-term fix for this would of course be for me to actually learn Rust and
its packaging system, and then update all of the dependencies to the latest
versions (dealing with and fixing compiler errors/warnings as I go) and then
sending it as a pull request to the core-rs repository on GitHub.

But for now, this is really all I can do. :(

The pkgrel has been bumped to 5 for this change, as it results in a different
binary then previously.