Remove unneeded workaround and fix build
Something was changed in the build process which caused make build to
fail when building the bindata target.
Explicitly setting the tags for both make generate and build fixes that.
Also, someone apparently pushed the tag v1.12.0-dev again months after
the branch was created, so the package version is now lower than before,
with only 15 commits since the tag as opposed to 37 before.