chore(weechat-notify-send): remove msg2 invocations
This patch removes the calls to the internal function `msg2`, moving the
post-installation instructions and informational messages (e.g.
"Installing the plugin...") to comments within the PKGBUILD. This
results in less output when building the package.

The post-installation instructions are helpful to display, but are out
of place in various scenarios, such as when the user builds the package
in a chroot environment, or as a different user than their own on the
host machine.