upgpkg: zfs-utils-git 0.8.0.r709.g5a42ef04f-1
This package is atrocious and completely unacceptable as-is. It claims
to be a git package, but uses pinned sources, and the packager has
simply ignored this for 15 months now.

The version number is honestly bonkers, it distributes tons of garbage,
and it's full of pseudocode from archzfs's questionable maintenance
scripts. Nuke it all and replace it with parallel code from zfs-dkms.