feat(google-cloud-sdk): only apply patches which are present in $source
This patch refactors the block within the prepare() function which
handles applying patches to explicitly use the patch files present in
the source array. Previously, the iteratable list was generated using a
shell glob -- doing this allows for a potential issue.

The particular issue comes into play when a user has
built a version of the package which contains, for example, three patch
files named 0{1-A,2-B,3-C}.patch -- when building manually with
`makepkg`, their local `src/` directory will contain all three patch
files. If a future version of this package removed the need for
03-C.patch, and the user built the package via `makepkg` again _without
passing the `--cleanbuild` flag__, the 03-C.patch file would not be
removed, and the shell globbing would detect and attempt to apply that
file, causing an error (assuming the patch failed to apply).

What we're doing now is much more sane, without introducing additional
overhead: iterating over each entry in $source and pattern-matching the
end of the string to ensure it ends with `\.patch` before applying it
allows us to add and remove patch files from $source as needed without
having to update prepare(), with the added benefit of bypassing the
issue referenced above.

closes #20