Add pkgrel to downloaded source name to sidestep caching issues
Upstream keeps changing the zip without changing the name or even the
version of the font inside, so people are having trouble building this
because they 'have' the download but their checksums fail. This forces
a new download filename every time we bump the package in the AUR, hence
giving users a much better chance of having the right file, not just
a file of the same name.