git-extras*/: make sure the PKGBUILD runs non-interactively
The Makefile has a "feature" where it checks the user's gitconfig to see
if there are aliases which match installed commands. If there are, it
will interactively prompt the user to decide whether to install them
anyway (defaulting to "yes").

This is a package, so obviously we want to install it. But a PKGBUILD is
supposed to run non-interactively, so prevent the question from even
being raised. We do this by setting the GIT_CONFIG variable to
/dev/null, preventing any other gitconfig files from being read.