Fix patch version for the current tag version
Patch version should be 'z' when version is in the format
'x.y.z'. When version is in the format 'x.y' it should be
'0'.

We can automate this in prepare() if upstream keeps changing
the format of the tags between 'x.y.z' and 'x.y' all the time.