feat(copybara-git): update pkgver to track revision count
As of this writing, there are not tags in the [upstream repository][0],
which makes determining a meaningful package version difficult, as a
pkgver comparison will simply use the ASCII key value for characters in
the 8-digit reference that is currently used (generated via `git-log`).

This patch solves this problem by counting the total number of revisions
and consuming that in the `pkgver`; this should help downstream
consumers be able to determine whether or not a revision is newer than
what they currently have. To note, the format of this pkgver is not
dissimilar from what we typically see in projects with tags (from
`git-describe`); we simply have to do an extra step here.