Make video settings and controller configuration windows to work
It looks like that mupen64plus-gui is part of a broader project
named m64p. It does not work as intended when used standalone
because it needs special compoenents, as follows.

The video settings window requires a patched/custom version
of GLideN64 that is made specially for mupen64plus-gui/m64p.
This special version of GLideN64 is mandatory in this case.

The controller configuration window requires the
mupen64plus-input-qt plugin.

So it looks logical to ship these mandatory components with
mupen64gui-git so it can work properly. Note that now the
package will conflict with the regular GLideN64 plugin and
m64p.