Update to 0.2.1.r452.1f28294
There have been many updates to the OCP git version. I updated the
PKGBUILD file again to work with the latest version.

Unfortunately, the MAD support had to be left out for now as there were
missing build instructions for charset.o. I'll try to introduce it back
again with the new revisions.

The package ttf-unifont is now a required dependency. A new patch was
needed that changes the font filenames to match the package.