Make some small tweaks to the Heroku package (#2)
* Do not show a progress bar when packaging

Disable the progress bar during dependency installation. The progress bar isn't necessary when creating a package and has been the source of significant slowdown in the past.

* Fix warnings about reference to srcdir/pkgdir

Remove all references to the $srcdir and $pkgdir directories in the packaged files. These are all internal references for Node that aren't necessary so we shouldn't leave them hanging behind.

* Use a local cache directory instead of $HOME

Use a local cache directory for NPM to prevent pollution of the builder's $HOME directory.