Initial version.
Installed as per instructions from the manual instal package from the
source web site.

Couldn't just add path to bin folder as the application appears to need
to be launched from the `/usr/local/AstroDMx_Capture/bin` so I created
and alias `astrodmx` that changes to that folder, runs and then returns
to starting folder.