Build with GTK+3 instead of GTK+2
I'm not sure if this functionality was present in v3.2 or if it's a new
thing with v3.3, but amethyst can now be built using GTK+3 by passing
the argument "hiro=gtk3" to make. So we do that here.

This has the side effect of removing all compiler warnings from the
previous commit, as we are no longer using any deprecated APIs!

However, it does cause some theme issues to occur. When no file is
open, the panel where the source editor would be becomes the system
theme's background colour rather than the one configured in
settings.bml. The same occurs in the unoccupied space in the file tree
panel. I'll see if I can fix this with a patch shortly.