Add patch to disable widget theming
amethyst uses the hiro UI abstraction library also used in higan, bsnes
etc. hiro supports GTK+3, but with a significant problem - many (not
all, but some) of the theme colours specified in settings.bml are
overridden by any CSS theme you have installed on your system (Adwaita,
Breeze, etc.). All of the APIs that hiro uses to perform colour
adjustment are from GTK+2, and are now deprecated. The proper solution
to this is to revamp hiro to use CSS for colour theming rather than the
outdated API. But this is something that should be done in a fork of
hiro so that the fix can be pulled into higan and bsnes as well. So for
now, I just comment out the calls to hiro's setBackgroundColor() and
setForegroundColor() functions, so that all UI widgets just use the
system CSS theme.