Keep the bin folder (by renaming glslc)
Since there is no distinct package to separate the libraries from the
executables, both should be kept in case someone should need any of them.

Use the same naming scheme as found in lib32-shaderc-git (there is no
naming convention for 32 bit executables on 64 bit architecture
to the best of my knowledges and searches).

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>