Fix build: issue with submodule mingw-std-threads
Submodule mingw-std-threads is targeting an untracked commit
(or something of that sort).

The target commit cannot be checked out by makepkg sources
(makepkg gives an error when using '#commit=<hash>'), but
it's available for fetching at upstream, so I'm manually
fetching it at prepare().

It seems that the target commit was somehow moved by upstream
mingw-std-threads, because it was working previously and now
it stopped to work. Maybe they deleted it from its branch.