Add missing library file dependencies
* upgpkg: sile-git 0.10.4.r17.g2c879c9-1
* upgpkg: sile 0.10.4-3

Per Arch wiki, use find-libdeps to scout out libraries we are
dynamically linked to after building. Some of the resulting list are not
properly provided upstream, but for the ones that are add these as
dependencies so that they will be exact version matched.

This issue is particularly noticeable with ICU and Harfbuzz which update
more frequently. When they do the sile packages break and need
rebuilding. This should help flag that as it happens so the breakage
isn't unexpected (i.e. throw error during system upgrade rather than
leaving sile to crashing on the next invocation).