Workaround issue with GCC 10
GCC 10 defaults to `-fno-common`¹, producing a linker error with current
codebase.  As a workaround, I've forced `-fcommon` when configuring.  See also
the topic² on bbs.archlinux.org for more details.

Also, it seems that the `validpgpkeys` entry was not added in `.SRCINFO`,
possibly because I used `mksrcinfo` to create it.  Now that I'm using `makepkg
--printsrcinfo`, it gets added nicely.

¹: https://gcc.gnu.org/gcc-10/porting_to.html
²: https://bbs.archlinux.org/viewtopic.php?id=255727