Remove explicit version dependency for cloud sdk components.
If require a specific version of the google-cloud-sdk package in one of
the extra components, then this will make it difficult (impossible?) to
perform an in-place upgrade of these packages (since you can not upgrade
the google-cloud-sdk package without breaking the existing install of
the google-cloud-sdk-datastore-emulator package, and you can not build
the new google-cloud-sdk-datastore-emulator package without the new
google-cloud-sdk package being installed).

I think the only way to do this where all of the parts have the required
version would be to use a split package. But I suspect most people don't
want to have to download the datastore/app engine parts.

I suspect the next best thing is to just have a requirement on the
correct package, and assume that people will upgrade all of them at the
same time (and if they don't, they get to keep the broken pieces).