Update to 2020 release
It is May of 2020 and langkit is still based entirely on python 2.
Because many of the required makedepends have been dropped from the
repos, they would have to be recreated in the AUR. Nobody got time for
that, so we simply download all the required wheels and install them in
a virtualenv with pip.