Optionally use custom rebase of xwayland_hidpi after upstream XWayland split.
Users just have to toggle the use of the upstream PR and use my branch
instead. It’s a custom branch, so I made it optional.
I’ve also bumped the version number. Not bumping it for the previous
commit may have caused problems to some users for a short amount of
time.