Bump binutils version to 2.34.
There is a breaking change in API used by hsdis: to be able to handle a
corner case, the bfd_octets_per_byte function now accepts one more argument
(the section). Given that hsdis uses only a fake binfd to obtain some
machine-specific parameters, there is no need in this parameter, and it
can just be NULLed.