Actually, swap back over to the official repository.
I don't know what the guidelines of the AUR are, but I don't want
to risk having my privileges revoked for not sticking to official
repositories.

My personal take on using my personal repository originally was because
the repository hasn't seen that much activity since January and the program
has been in a light maintenance state for a while. I might consider creating
an alternative cmus-git package though with my changes.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>