use the launcher
It doesn't make sense to not use the launcher for general consumption
because the RuneLite client keeps the vanilla game up to date, but
cannot keep itself up to date. If it gets out of sync with it's the game
it breaks or enters a heavily reduced functionality mode. The launcher
ensures these are in sync properly and is the standard way for us to
release timely updates to the entire userbase.