zfs-dkms*/: fix configure.ac modifications to allow all module/ files
We want the toplevel Makefile/zfs.release and we *also* want all files
in module/ to be configured, because that is the only subdirectory we
use. It just happened that module/ used to only have Makefile.in to
configure, so only permitting Makefiles which may or may not be prefixed
with 'module/' worked.

The current upstream master now has a Kbuild.in file, though, which
broke this clumsy hack. Update to use a more beautiful variant of clumsy
hack instead.