Revert "new maintainer"
This reverts commit 2ba3bdacce6da1243b443193941643401c44d038.

It's completely invalid, some drive-by "maintainer" adopting a package
only to break a working pkgver() function into something that does not
work, and not adding anything else whatsoever. The whole commit needs to
go, and it would be better if it had never existed but the AUR doesn't
allow force pushing to revert to old commits.