Use pinned commit frag
Installing this package results in complaints being logged to dmesg.  These
may be the result of a conflict between

  * cros_kbd_led_backlight.ko (shipped with linux)
  * chromeos_keyboard_leds.ko (ours)

both of which, for example, claim the alias platform:chromeos-keyboard-leds.
Some workarounds include

   * blacklisting the existing "cros" module via modprobe.d
   * manually loading ours via modules-load.d
   * renaming the existing file to cros_kbd_led_backlight.ko.xz.orig

Of these, the second is chosen because upstream mentions it in its README.

A change to the core linux package on 2019-10-14 has removed auto-creation of
the main extramodules dir and symlinks to it from children of versioned
siblings.  Not sure whether this was in response to a shift in upstream
behavior, but module discovery appears unchanged.

See: 1fef2fd41ff7ad7f8713295471ee44adb9029c86

Although real archlinux packages like nvidia still use the extramodules
convention, this package has reverted to traditional upstream behavior.

See: https://git.archlinux.org/svntogit/packages.git
     /tree/trunk/PKGBUILD?h=packages/nvidia#n27