Fix package version
Technically, version 3.0 is in development at the time of writing, but
the earlier 2.*-tags are not reachable anymore.  Since `git describe`
won't work, we fall back to counting revisions.  This is forward
compatible with future versions `v3.*.*`.

Hopefully no one notices that this version change is not backwards
compatible, so I won't have to bump the package epoch literally 5
minutes after creating the package.