Fix license warning + keep bin content
This package is not a library only package. It contains a executable file
under the bin folder. Thus, we should keep it, with the proper suffix.
Otherwise, we would need to rename the package to lib32-libdav1d to make
the distinction, stating that it provides lib32-dav1d. Doing so would only
be partially true though, since it would be a subset of the full content.

Some other distros choose to provide splitted packages to distinguish
between libraries and executables / programs.

When/if the package is splitted in the future to propose a library and an
executable, we shall mirror this new reality.

Fix a namcap warning: since the system's architecture package is required,
create symlinks to the shared content. (Then again, this wouldn't be needed
if the content was splitted as done in other distros, allowing to use the
same content where needed).

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>