Split libraries from executables, rebuild against latest x265 and aom
Most peolple don't need the executables. Let's split libraries from the
executables in two different packages as done in other distro. If needed,
one can install lib32-ffmpeg to mimic exactly what is done by the ffmpeg
package.

Rebuild against latest x265 and aom.

We are not moving to 4.3 yet, since a major bug was reported with chromium
and others. Until it is fixed, we'll stay under the 4.2 branch.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>