bump to curl 7.71.0 alongside some build changes
As it turns out, it is assumed, that the user takes the current
master of quiche of the release date of curl.

This has now been implemented by pinning the commit ref of
the quiche master on the release date. This is intended to get
updated on every update to this package.

In addition, this package is now built in a reproducible way :)