Bring back svt-av1. Temporarily disable srt (broken by upstream).
Upstream svt-av1 has provided support for the recent API
changes of ffmpeg git master[1], so it can be added again.

Upstream srt recently introduced a commit[2] that removes
deprecated API and breaks ffmpeg support. Bug reports were
opened upstream at srt[3] and ffmpeg[4], and it's up to
ffmpeg to fix it. Unfortunatelly it was marked as a 'wish'
at the ffmpeg bug tracker, so it can take very long to be
fixed.

References
----------
[1] https://github.com/OpenVisualCloud/SVT-AV1/commit/b255a71cf47038ec73873ae507a56bc9fdef7c8f
[2] https://github.com/Haivision/srt/commit/0e2201aff6b379979cec43fee5e8f162717f6346
[3] https://github.com/Haivision/srt/issues/1382
[4] https://trac.ffmpeg.org/ticket/8760