feat(terragrunt): set GOPATH to a path within $srcdir
This patch sets GOPATH to a path within $srcdir, which helps to
"sandbox" the build of this package for users invoking `makepkg`
directly or through the use of an AUR helper (instead of in a chroot or
real container).

We avoid cleaning this cache after builds so that A) future builds can
re-use the cache, B) we don't remove the compiled executable, and C) it
is not the responsibility of this package nor is it an expectation of
`makepkg` that there are no side-effects (files left over on disk).

closes #30