feat(terragrunt): allow module cache to be deleted
In commit ace59782f4084933782791b445e318bf141851c0, we avoided cleaning
the build cache beacuse we wanted to enable future builds to utilize the
cache, however, this proves problematic for users who rebuild the
package using `makepkg --clean`, as they are greeted with many lines
denying permission to remove the file(s). This is due to the mode set on
modules fetched by the compiler, which are set to 444.

This patch introduces a workaround: by changing the file mode of the cache, we
can benefit from reuse as well as allow users using `--clean` or `--cleanbuild`
(or other processes) to be able to remove it.