ground work for aarch64 (arm64) builds
Not much to change, only a newer version of the udev binding crate is
required, as else we get a build error due to a type error (char ==
u8 on one and char == i8 on the other platform).

A successful build is still blocked by a similar issue in our own
proxmox-backup crate, but that should be easy to address.

Signed-off-by: Thomas Lamprecht <thomas@lamprecht.org>