tetrio.desktop: rename to workaround xdg-settings bug
When opening TETR.IO, the default text/html handler is set to this app,
which is wrong. This seems to be an issue with xdg-settings handling
desktop files with - in their name wrong[0].

Workaround this bug by removing the -desktop prefix for the desktop
file.

I wasn't able to reproduce the bug but let's change it anyway because
clearly it's affecting _some_ people.

[0]: https://github.com/electron/electron/issues/20382

Reported-by: 1lann