Use shorter patch to address no-compile.patch
When building the package as one typically would via `makepkg`, no
errors are encountered. However, when building in a clean chroot with
the help of the `devtools` package, "no-compile.patch" cannot be found.
Addressing the file as one directory "closer" to the current working
directory solves this problem, and `makepkg` still finds the file due to
a symlink it creates automatically.

This will be useful for both testing the package and moving it to the
community repository, should that ever happen.