Fix Version Mismatch In Version Patch File
The `version-fix.patch` file did not get updated to `0.9.2` when the rest of the package was. Since metals does some runtime resolution of artifacts, in particular in this case `mtags`, this caused metals to fail.

This commit corrects the issue. In the future, since upstream doesn't set this value in release, we may want to dynamically generate the patch file (sounds gross, but relying on human memory to do this is unreliable).