Normalize commit abbreviated hash in pkgver to only 7 chars
While 12-char hashes are desirable for more sensitive applications (and
users should gradually move towards that), 7-char hashes are fine for
the package version.  More importantly, the popular packages in the AUR
use the short/old hash sizes.

Normalize ours to that size and override the user's (or my one) global
git configuration.